unify default tools loading
