+++ tests/README.md
@@ -6,6 +6,11 @@ This directory contains **unit** and **integration** tests for the `actors-mcp-s
 
 Unit tests are located in the `tests/unit` directory.
 
+To run the unit tests, you can use the following command:
+```bash
+npm run test:unit
+```
+
 # Integration Tests
 
 Integration tests are located in the `tests/integration` directory.
@@ -16,4 +21,9 @@ ALL DEFAULT ONES DEFINED IN consts.ts AND ALSO EXPLICITLY:
 apify/rag-web-browser
 apify/instagram-scraper
 apify/python-example
+```
+
+To run the integration tests, you can use the following command:
+```bash
+APIFY_TOKEN=your_token npm run test:integration
 ```
